+++ tests/test_libre.py
@@ -6,27 +6,6 @@ import pytest
 from deep_translator import exceptions, LibreTranslator
 from deep_translator.constants import LIBRE_CODES_TO_LANGUAGES
 
-test_text_standard = 'Hello world.'
-TRANSLATED_RESULTS = {
-    'English': 'Hello world.',
-    'Arabic': 'مرحبا العالم.',
-    'Chinese': '霍洛美世界。',
-    'French': 'Bonjour.',
-    'German': 'Hallo Welt.',
-    'Hindi': 'नमस्ते दुनिया।',
-    'Indonesian': 'Halo dunia.',
-    'Irish': 'Dia duit domhan.',
-    'Italian': 'Ciao mondo.',
-    'Japanese': 'こんにちは。',
-    'Korean': '안녕하세요.',
-    'Polish': 'Hello world (ang.).',
-    'Portuguese': 'Olá, mundo.',
-    'Russian': 'Привет мир.',
-    'Spanish': 'Hola mundo.',
-    'Turkish': 'Merhaba dünya.',
-    'Vietnamese': 'Xin chào.'
-}
-
 
 @pytest.fixture
 def libre():
@@ -35,31 +14,28 @@ def libre():
 
 def test_content(libre):
     """Sample pytest test function with the pytest fixture as an argument."""
-    # from bs4 import BeautifulSoup
-    # assert 'GitHub' in BeautifulSoup(response.content).title.string
     assert libre.translate(text='good') is not None
 
 
 def test_inputs():
-    with pytest.raises(exceptions.LanguageNotSupportedException):
+    with pytest.raises(exceptions.InvalidSourceOrTargetLanguage):
         LibreTranslator(source="", target="")
 
-    with pytest.raises(exceptions.LanguageNotSupportedException):
-        LibreTranslator(source="auto", target="nothing")
+    with pytest.raises(exceptions.InvalidSourceOrTargetLanguage):
+        LibreTranslator(source="auto", target="")
+
+    with pytest.raises(exceptions.InvalidSourceOrTargetLanguage):
+        LibreTranslator(source="", target="en")
 
 
 def test_abbreviations_and_languages_mapping():
     for abb, lang in LIBRE_CODES_TO_LANGUAGES.items():
-        if abb != 'en':
-            l1 = LibreTranslator(abb)
-            l2 = LibreTranslator(lang)
-            assert l1.source == l2.source
+        l1 = LibreTranslator(abb)
+        l2 = LibreTranslator(lang)
+        assert l1._source == l2._source
 
 
 def test_payload(libre):
-    with pytest.raises(exceptions.NotValidPayload):
-        libre.translate("")
-
     with pytest.raises(exceptions.NotValidPayload):
         libre.translate(123)
